# docs/swagger.yaml
basePath: /
definitions:
...
  dto.SubscriptionPlanResponse:
    properties:
      currency:
        type: string
      duration:
        type: string
      features:
        additionalProperties: true
        type: object
      id:
        type: string
      limits:
        additionalProperties: true
        type: object
      name:
        type: string
      price:
        type: number
    type: object
...
  dto.UserSubscriptionResponse:
    properties:
      auto_renew:
        type: boolean
      end_date:
        type: string
      id:
        type: string
      plan_id:
        type: string
      start_date:
        type: string
      status:
        type: string
      usage:
        additionalProperties: true
        type: object
      user_id:
        type: string
    type: object
...
  handlers.SubscribeRequest:
    properties:
      auto_renew:
        type: boolean
      plan_id:
        type: string
      user_id:
        type: string
    type: object
...
  models.CastingResponse:
    properties:
      castingID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      modelID:
        type: string
      status:
        description: '"pending" | "accepted" | "rejected"'
        type: string
    type: object
  models.EmployerProfile:
    properties:
      city:
        type: string
      companyName:
        type: string
      companyType:
        type: string
      contactPerson:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      isVerified:
        type: boolean
      phone:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
      website:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
    type: object
  models.ModelProfile:
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Модельное агенство
    url: http://contact.company.com
  description: API для управления модельной компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: mwork API
  version: "1.0"
paths:
...
  /employer-profiles:
    post:
      consumes:
      - application/json
      description: Создаёт новый профиль работодателя на основе переданных данных
      parameters:
      - description: Данные профиля работодателя
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.EmployerProfile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EmployerProfile'
        "400":
          description: Некорректный JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при создании профиля
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать профиль работодателя
      tags:
      - employer-profiles
  /employer-profiles/{user_id}:
    get:
      consumes:
      - application/json
      description: Возвращает профиль работодателя, связанный с указанным user_id
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmployerProfile'
        "404":
          description: Профиль не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить профиль работодателя по ID пользователя
      tags:
      - employer-profiles
  /model-profiles:
    post:
      consumes:
      - application/json
      description: Создаёт новый профиль модели на основе переданных данных
      parameters:
      - description: Данные модельного профиля
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.ModelProfile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ModelProfile'
        "400":
          description: Некорректный JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при создании профиля
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать модельный профиль
      tags:
      - model-profiles
  /model-profiles/{user_id}:
    get:
      consumes:
      - application/json
      description: Возвращает профиль модели, связанный с указанным user_id
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModelProfile'
        "404":
          description: Профиль не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить профиль модели по ID пользователя
      tags:
      - model-profiles
  /responses:
    get:
      consumes:
      - application/json
      description: Возвращает список откликов по ID кастинга
      parameters:
      - description: ID кастинга
        in: query
        name: casting_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CastingResponse'
            type: array
        "500":
          description: Ошибка при получении откликов
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить отклики на кастинг
      tags:
      - responses
    post:
      consumes:
      - application/json
      description: Создаёт отклик на кастинг со статусом "pending"
      parameters:
      - description: Отклик на кастинг
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/models.CastingResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CastingResponse'
        "400":
          description: Некорректный JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при создании отклика
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Отклик на кастинг
      tags:
      - responses
  /responses/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает отклик по ID
      parameters:
      - description: ID отклика
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CastingResponse'
        "404":
          description: Отклик не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить отклик по ID
      tags:
      - responses
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку на основе переданных данных
      parameters:
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SubscribeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserSubscriptionResponse'
        "400":
          description: Некорректный ввод
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при создании подписки
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Оформить подписку
      tags:
      - subscriptions
  /subscriptions/plans:
    get:
      consumes:
      - application/json
      description: Возвращает все доступные планы подписки
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionPlanResponse'
            type: array
        "500":
          description: Ошибка при получении планов
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить список подписочных планов
      tags:
      - subscriptions
  /subscriptions/usage-limit:
    get:
      consumes:
      - application/json
      description: Проверяет, можно ли выполнить указанное действие в рамках текущей
        подписки
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: 'Тип действия (например: upload, apply)'
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'allowed: bool, remaining: int'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Отсутствует user_id или action
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при проверке лимита
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Проверить лимит использования
      tags:
      - subscriptions
  /subscriptions/user/{userID}:
    get:
      consumes:
      - application/json
      description: Возвращает текущую активную подписку по ID пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserSubscriptionResponse'
        "400":
          description: userID не указан
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить активную подписку пользователя
      tags:
      - subscriptions
...
swagger: "2.0"